requirements.txt
Django==2.2.19
Pillow==9.1.1
python-dotenv==0.20.0
pytz==2022.1
sorl-thumbnail==12.8.0
sqlparse==0.4.2
